test(PDO): move parseDsn quote-handling cases into PDOTest
